[FIX] pms_w_s: tests
Using URLLookup `model()` does not work for public user as public user
has no specific rights on any object.

Another way to solve it is to give read rights for pblic user. But that
seams dangerous to me as any request made on the website is performed by
public user.