[FIX] pms_w_s: fix price for rooms
Using `price_total` from the `availability_result` does not work as by
default in the booking.engine wizard the number of selected rooms
(`num_room_selected`) is equal to zero. Therefor the `price_total` is
null.

To solve that we set the amount of room selected to 1 for each rooms
available.

As the booking.engine for the rooms page is not stored this is a
temporary value that will not be used for booking.